+++ qe-preview/CHANGELOG.md
@@ -6,4 +6,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+## [0.0.3] - 2024-04-23
+### Added
+- Added support for customizing the regex matching rules via Edit Settings.
+
+## [0.0.2] - 2024-04-14
+### Added
+- Added support for customizing the regex matching rules using a JSON configuration file.
+- Added a screenshot to the README.md file.
+- Added *click to copy* feature for the output data.
+